Move DirectoryHTMLBuilder to sphinx.builders.dirhtml package
